vur_api_project/requirements.txt
attrs==22.2.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cloud-init==19.1.23
configobj==5.0.8
cryptography==40.0.2
databricks-sql-connector==1.0.2
dataclasses==0.8
DBUtils==3.0.3
distro==1.8.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.0.3
Flask-Cors==4.0.0
future==0.18.3
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.3
jsonschema==3.2.0
kafka-python==2.0.2
MarkupSafe==2.0.1
oauthlib==3.2.2
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.0
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
texttable==1.7.0
thrift==0.16.0
typing_extensions==4.1.1
urllib3==1.26.18
websocket-client==0.59.0
Werkzeug==2.0.3
zipp==3.6.0
